add all services by Dr.Martin (Martin Java)

+++ src/main/java/com/example/demo/repository/RoomRepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepo extends JpaRepository<Room,Integer> {
+public interface RoomRepo extends JpaRepository<Room,Long> {
 }
